--- RevenueCat/Scripts/PaywallResult.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a78f15c4e2b3e4e2ba4f5a67e8b21b23
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
