+++ RevenueCat/Scripts/Purchases.cs
@@ -608,8 +608,6 @@ public partial class Purchases : MonoBehaviour
         _wrapper.SetLogHandler();
     }
 
-    private System.Action<PaywallResult> _oneShotPaywallResultCallback;
-
     private CustomerInfoFunc GetCustomerInfoCallback { get; set; }
 
     // ReSharper disable once UnusedMember.Global
@@ -1502,21 +1500,6 @@ public partial class Purchases : MonoBehaviour
     }
 
     // ReSharper disable once UnusedMember.Local
-    private void _paywallResult(string paywallResultName)
-    {
-        Debug.Log("_paywallResult " + paywallResultName);
-        
-        var callback = _oneShotPaywallResultCallback;
-        _oneShotPaywallResultCallback = null;
-        
-        if (callback != null)
-        {
-            var result = PaywallResultExtensions.FromString(paywallResultName);
-            callback(result);
-        }
-    }
-
-    // ReSharper disable once UnusedMember.Local
     private void _restorePurchases(string customerInfoJson)
     {
         Debug.Log("_restorePurchases " + customerInfoJson);
@@ -1894,12 +1877,5 @@ public partial class Purchases : MonoBehaviour
     private static bool ResponseHasError(JSONNode response)
     {
         return response != null && response.HasKey("error") && response["error"] != null && !response["error"].IsNull;
-    }
-
-    // Present paywall and receive a PaywallResult
-    public void PresentPaywall(System.Action<PaywallResult> callback)
-    {
-        _oneShotPaywallResultCallback = callback;
-        _wrapper.PresentPaywall(callback);
     }
 }
