--- RevenueCatUI/Editor/RevenueCat.UI.Editor.asmdef.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e6c6f27f8e4c4f6e9e0d8a9a8f1b2c3d
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-
