--- RevenueCatUI/Editor/RevenueCatUI.Editor.asmdef
@@ -1,19 +1,0 @@
-{
-    "name": "RevenueCatUI.Editor",
-    "rootNamespace": "RevenueCat.UI",
-    "references": [
-        "RevenueCatUI"
-    ],
-    "includePlatforms": [
-        "Editor"
-    ],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
-
