+++ RevenueCatUI/Plugins/Android/AndroidManifest.xml
