+++ RevenueCatUI/Plugins/Android/AndroidManifest.xml.meta
