+++ RevenueCatUI/Plugins/Android/RevenueCatUI.java
@@ -1,6 +1,9 @@
 package com.revenuecat.unity.ui;
 
 public class RevenueCatUI {
+    
+    private static final String PAYWALL_EVENT = "_paywallEvent";
+    private static final String PAYWALL_RESULT = "_paywallResult";
 
     // Java-side callbacks registered from C#; no UnitySendMessage fallback.
     public interface PaywallCallbacks { void onPaywallResult(String result); }
@@ -10,10 +13,24 @@ public class RevenueCatUI {
     public static void registerPaywallCallbacks(PaywallCallbacks cb) { paywallCallbacks = cb; }
     public static void unregisterPaywallCallbacks() { paywallCallbacks = null; }
 
-
     public static void presentPaywall(String offeringIdentifier) {
-        android.util.Log.d("RevenueCatUI", "presentPaywall(offering=" + offeringIdentifier + ")");
+        android.app.Activity activity = UnityPlayer.currentActivity;
         sendPaywallResult("CANCELLED|Stub: no native UI");
+    }
+
+    public static void presentPaywallForResult(@Nullable String offeringIdentifier) {
+        try {
+            android.app.Activity activity = UnityPlayer.currentActivity;
+            Intent intent = new Intent(activity, com.revenuecat.purchasesunity.PaywallProxyActivity.class);
+            intent.putExtra(com.revenuecat.purchasesunity.PaywallProxyActivity.EXTRA_GAME_OBJECT, gameObject);
+            intent.putExtra(com.revenuecat.purchasesunity.PaywallProxyActivity.EXTRA_METHOD, PAYWALL_RESULT);
+            if (offeringIdentifier != null) {
+                intent.putExtra(com.revenuecat.purchasesunity.PaywallProxyActivity.EXTRA_OFFERING_ID, offeringIdentifier);
+            }
+            activity.startActivity(intent);
+        } catch (Throwable t) {
+            Log.e("Purchases", "Error launching PaywallProxyActivity", t);
+        }
     }
 
     public static void presentPaywallIfNeeded(String requiredEntitlementIdentifier, String offeringIdentifier) {
