+++ RevenueCatUI/Plugins/Editor/RevenueCatUIDependencies.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<dependencies>
+  <androidPackages>
+    <androidPackage spec="com.revenuecat.purchases:purchases-hybrid-common-ui:[17.7.0]" />
+    <androidPackage spec="androidx.appcompat:appcompat:1.6.1" />
+  </androidPackages>
+  
+  <iosPods>
+    <iosPod name="PurchasesHybridCommonUI" version="~> 17.7.0" />
+  </iosPods>
+</dependencies>
