--- RevenueCatUI/Plugins/iOS/RevenueCatUI.m.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6f596540f84544a1582c80e4f69a709f
