--- RevenueCatUI/Scripts/Platforms/iOS.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 49452c74706dd4f08a6c9af8de46966e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
