--- RevenueCatUI/Scripts/Platforms/iOS/IOSPaywallPresenter.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 3f43077ff755543a5b17a0b90808573d
-
